Moved Header and Footer to componens directory
