+++ src/stories/components/Header.svelte
@@ -5,7 +5,7 @@
 	import { Button } from '$lib/components/ui/button';
 
 	import './header.css';
-	import logo from './assets/logo.png';
+	import logo from '../assets/logo.png';
 	import { fn } from '@storybook/test';
 
 	interface Props {
